requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
langchain==0.0.340
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-docx==0.8.11
unstructured==0.10.30
beautifulsoup4==4.12.2
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27
